Added reading of IMU and FT port
Local ports are open during device opening phase.
They are connected to the remote ports during the attach.

The configuration has the following form:
```xml
    <paramlist name="KEY">
        <elem name="imu_id">remote_port</elem>
    </paramlist>
```
where KEY = "imus" for the IMUs section and "fts" for the FTs section.

Ports are read asynchronously by callbacks.
Data is saved in a buffer which is then read by the run() method of the bridge and serialised on the network.